Move DominantResourceShare methods to more appropriate files
